requirements.txt
aiofiles==23.2.1
aiohttp==3.8.3
aiosignal==1.2.0
airtest==1.2.7
allure-pytest==2.13.2
allure-python-commons==2.13.2
altair==5.1.2
altgraph==0.17.4
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
Appium-Python-Client==3.1.0
APScheduler==3.10.1
arrow==1.3.0
asgiref==3.5.2
async==0.6.2
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
autopep8==1.7.0
Babel==2.13.1
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.92
bcrypt==4.0.1
BeautifulReport==0.1.3
beautifulsoup4==4.7.1
bidict==0.22.0
black==22.12.0
blinker==1.6.3
Brotli==1.0.9
cached-property==1.5.2
certifi==2022.6.15.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
colorlog==4.8.0
comtypes==1.1.14
ConfigArgParse==1.5.3
contourpy==1.1.1
cryptography==38.0.4
cycler==0.12.1
dataclasses-json==0.6.1
datasets==2.14.6
ddddocr==1.4.8
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.4
distro==1.8.0
Django==4.1.3
django-apscheduler==0.6.2
django-mysql==4.7.1
django-redis==5.2.0
django-simpleui==2022.12.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==4.4.4
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
drf-extensions==0.7.1
drf-spectacular==0.24.2
dubboclient==1.0.1
easydict==1.11
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.0.4
execnet==2.0.2
facebook-wda==1.4.6
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
Flask==2.2.2
flask-babel==4.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.3.1
fsspec==2023.10.0
func-timeout==4.3.5
funppy==0.5.2
future==0.18.3
gevent==22.10.2
geventhttpclient==2.0.8
gitdb2==2.0.6
GitPython==2.1.15
gradio==4.1.1
gradio_client==0.7.0
greenlet==3.0.1
grpcio==1.59.2
grpcio-tools==1.59.2
h11==0.14.0
har2case==0.3.1
hrpc==1.0.9
httpcore==0.18.0
httprunner==2.5.7
httpx==0.25.0
huggingface-hub==0.18.0
humanfriendly==10.0
idna==2.8
importlib-metadata==5.0.0
importlib-resources==6.1.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==2.10.3
jmespath==0.9.5
joblib==1.3.2
jsonpatch==1.33
jsonpath==0.82.2
jsonpointer==2.4
jsonschema==4.17.0
keyboard==0.13.5
kiwisolver==1.4.5
langchain==0.0.320
langsmith==0.0.49
locust==2.18.3
locustio==0.9.0
loguru==0.4.1
lxml==4.9.3
lyrebird==2.20.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.7.3
mdurl==0.1.2
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.4
mss==6.1.0
multidict==6.0.2
multiprocess==0.70.12.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
netifaces==0.11.0
nose==1.3.7
numpy==1.24.4
onnx==1.15.0
onnxruntime==1.16.1
openai==1.2.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
orjson==3.9.10
outcome==1.2.0
packaging==23.2
paddle2onnx==1.0.6
paddlefsl==1.1.0
paddlehub==2.4.0
paddlenlp==2.6.1
pandas==2.0.3
paramiko==2.12.0
pathspec==0.11.2
pefile==2023.2.7
Pillow==9.4.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
playwright==1.38.0
pluggy==1.0.0
pocoui==1.0.88
poium==1.3.2
portpicker==1.3.1
protobuf==3.20.2
psutil==5.9.4
py==1.11.0
pyaml==19.12.0
pyarrow==14.0.0
PyAutoGUI==0.9.54
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyee==9.0.4
PyGetWindow==0.0.9
PyGithub==1.45
Pygments==2.16.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyJWT==2.6.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyocr==0.8.5
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.2
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.0
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-playwright==0.4.3
pytest-rerunfailures==12.0
pytest-timeout==2.2.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.3.4
python-gitlab==1.15.0
python-hessian==1.1.2
python-jenkins-monitor==1.4.1
python-multipart==0.0.6
python-slugify==8.0.1
python-socketio==5.7.1
pytils==0.4.1
pytweening==1.0.7
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pywin32==227
pywin32-ctypes==0.2.2
pywinauto==0.6.3
PyYAML==5.3
pyzmq==25.0.0
qrcode==7.3.1
rarfile==4.1
redis==4.4.0
requests==2.25.1
requests-toolbelt==0.9.1
retry==0.9.2
rich==13.6.0
roundrobin==0.0.4
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.12.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==0.13.5
seqeval==1.2.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.22
sqlparse==0.4.3
starlette==0.27.0
svn==1.0.1
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
texttable==1.6.7
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
tools==0.1.9
toolz==0.12.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2022.7
tzlocal==4.2
ucimlrepo==0.0.3
uritemplate==4.1.1
urllib3==1.24.3
uvicorn==0.24.0
visualdl==2.5.3
websocket-client==0.48.0
websockets==11.0.3
Werkzeug==2.2.2
wfastcgi==3.0.0
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.2.0
XMind==1.2.0
xmind2testcase==1.5.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.8.1
zipp==3.10.0
zope.event==4.6
zope.interface==5.5.2
